update the base and utilities
