+++ molecularnodes/entities/utilities.py
@@ -1,0 +1,104 @@
+"""
+Python module for converting molecular structures from biotite into Blender objects.
+
+This module provides functions to:
+1. Create 3D models from biotite AtomArray objects (`create_object`)
+2. Attach molecular information as attributes to Blender objects (`atom_array_to_named_attributes`)
+
+The created Blender objects can be used for visualization and manipulation of molecular
+structures within Blender's 3D environment. Properties like atom positions, bonds, and
+other molecular attributes are preserved as named attributes that can be used with
+Geometry Nodes and other Blender features.
+"""
+
+import bpy
+import databpy
+from databpy import AttributeDomains, AttributeTypes
+from biotite.structure import AtomArray,  AtomArrayStack
+import numpy as np
+
+__all__ = ["create_object", "atom_array_to_named_attributes"]
+
+
+def create_object(
+    array: AtomArray,
+    name="NewObject",
+    centre: str | None = None,
+    world_scale: float = 0.01,
+    collection: bpy.types.Collection | None = None,
+) -> bpy.types.Object:
+    """
+    Create a 3D model of the molecule, with one vertex for each atom.
+
+    Parameters
+    ----------
+
+    Returns
+    -------
+    bpy.types.Object
+        The created 3D model, as an object in the 3D scene.
+    """
+    if isinstance(array, AtomArrayStack):
+        array = array[0]
+
+    bob = databpy.create_bob(
+        vertices=array.coord * world_scale,
+        edges=array.bonds.as_array()[:, :2] if array.bonds is not None else None,
+        name=name,
+        collection=collection,
+    )
+    # Add information about the bond types to the model on the edge domain
+    # Bond types: 'ANY' = 0, 'SINGLE' = 1, 'DOUBLE' = 2, 'TRIPLE' = 3, 'QUADRUPLE' = 4
+    # 'AROMATIC_SINGLE' = 5, 'AROMATIC_DOUBLE' = 6, 'AROMATIC_TRIPLE' = 7
+    # https://www.biotite-python.org/apidoc/biotite.structure.BondType.html#biotite.structure.BondType
+    if array.bonds:
+        bob.store_named_attribute(
+            array.bonds.as_array()[:, 2],
+            "bond_type",
+            domain=AttributeDomains.EDGE,
+            atype=AttributeTypes.INT,
+        )
+
+    atom_array_to_named_attributes(array, bob.object, world_scale=world_scale)
+
+    if centre is not None:
+        bob.position -= bob.centroid(centre)
+
+    return bob.object
+
+
+
+def atom_array_to_named_attributes(array, obj, world_scale=0.01):
+    """All annotations in the AtomArray are stored as attributes on the mesh."""
+
+    # don't need to add coordinates as those have been stored as `position` on the mesh
+    annotations_to_skip = ["coord", "hetero"]
+
+    for attr in array.get_annotation_categories():
+        if attr in annotations_to_skip:
+            continue
+
+        data = array.get_annotation(attr)  # type: ignore
+
+        if attr == "vdw_radii":
+            data *= world_scale
+
+        # geometry nodes doesn't support strings at the moment so we can only store the
+        # numeric and boolean attributes on the mesh. All string attributes should have
+        # already been converted to a corresponding numeric attribute during the
+        # reader process
+
+        if not (
+            np.issubdtype(data.dtype, np.number)
+            or np.issubdtype(data.dtype, np.bool_)
+        ):
+            continue
+
+        # the integer versions of strings have been added as annotations that just
+        # append `_int` onto the name so we don't overrite the original data but when
+        # storing on the meshh we can just remove the `_int`
+        databpy.store_named_attribute(
+            obj=obj,
+            data=data,
+            name=attr.replace("_int", ""),
+        )
